# Fewer SoALoad/Extract calls