--- a/src/containers/LinkedCells.h
+++ b/src/containers/LinkedCells.h
@@ -135,7 +135,7 @@ class LinkedCells : public ParticleContainer<Particle, ParticleCell> {
   template <class ParticleFunctor>
   void iteratePairwiseSoA2(ParticleFunctor *f, bool useNewton3 = true) {
 
-    fillSoAs(f);
+    loadSoAs(f);
 
     auto envTraversal = std::getenv("AUTOPAS_TRAVERSAL");
     if (useNewton3) {
@@ -217,18 +217,18 @@ class LinkedCells : public ParticleContainer<Particle, ParticleCell> {
 
  protected:
   /**
-   * object to manage the block of cells
+   * object to manage the block of cells.
    */
   CellBlock3D<ParticleCell> _cellBlock;
   // ThreeDimensionalCellHandler
 
   /**
-   * Iterate over all cells and load the data in the SoAs
+   * Iterate over all cells and load the data in the SoAs.
    * @tparam ParticleFunctor
    * @param functor
    */
   template <class ParticleFunctor>
-  void fillSoAs(ParticleFunctor &functor) {
+  void loadSoAs(ParticleFunctor &functor) {
 #ifdef AUTOPAS_OPENMP
     //TODO find a condition on when to use omp or when it is just overhead
 #pragma omp parallel for
@@ -239,7 +239,7 @@ class LinkedCells : public ParticleContainer<Particle, ParticleCell> {
   }
 
   /**
-   * Iterate over all cells and fetch the data from the SoAs
+   * Iterate over all cells and fetch the data from the SoAs.
    * @tparam ParticleFunctor
    * @param functor
    */