--- a/src/containers/LinkedCells.h
+++ b/src/containers/LinkedCells.h
@@ -228,7 +228,7 @@ class LinkedCells : public ParticleContainer<Particle, ParticleCell> {
    * @param functor
    */
   template <class ParticleFunctor>
-  void loadSoAs(ParticleFunctor &functor) {
+  void loadSoAs(ParticleFunctor *functor) {
 #ifdef AUTOPAS_OPENMP
     //TODO find a condition on when to use omp or when it is just overhead
 #pragma omp parallel for
@@ -244,7 +244,7 @@ class LinkedCells : public ParticleContainer<Particle, ParticleCell> {
    * @param functor
    */
   template <class ParticleFunctor>
-  void extractSoAs(ParticleFunctor &functor) {
+  void extractSoAs(ParticleFunctor *functor) {
 #ifdef AUTOPAS_OPENMP
     //TODO find a condition on when to use omp or when it is just overhead
 #pragma omp parallel for