--- a/tests/testAutopas/AoSvsSoATest.cpp
+++ b/tests/testAutopas/AoSvsSoATest.cpp
@@ -59,7 +59,7 @@ TEST_F(AoSvsSoATest, testAoSvsSoA) {
     cell.addParticle(p);
   }
 
-  ljFunctor.SoALoader(cell, &cell._particleSoABuffer);
+  ljFunctor.SoALoader(cell, cell._particleSoABuffer);
   start = std::chrono::high_resolution_clock::now();
   ljFunctor.SoAFunctor(cell._particleSoABuffer, cell._particleSoABuffer);
   stop = std::chrono::high_resolution_clock::now();
@@ -71,7 +71,7 @@ TEST_F(AoSvsSoATest, testAoSvsSoA) {
   // copy back to particle array
   particlesSoA.clear();
 
-  ljFunctor.SoAExtractor(&cell, &cell._particleSoABuffer);
+  ljFunctor.SoAExtractor(cell, cell._particleSoABuffer);
 
   //  ASSERT_EQ(particlesAoS.size(), particlesSoA.size());
   ASSERT_EQ(particlesAoS.size(), cell.numParticles());