Referencing #42 

This removes the `SoALoad/Extract` calls from the cell functor since it called `Load/Extract` for each pairwise interaction. The `Load/Extract` is now done in separate iterations in the `iteratePairwiseSoA` methods of the containers.

Speed improvement on a grid example:

Container | Iterations | \#Particles | grid spacing | cutoff | 
--------------|--------------|--------------|---------------|-----------
Linked-Cells | 10 | 60^3 | 1 | 2.5 | 

Time for one iteratePairwise call in s:

Traversal | Old | New | Speedup
-------------|------|--------|--------------
c08 | .96 | .84 | 1.14
sli | .93 | .82 | 1.13