@FG-TUM 2018-05-27
I think there is still room for improvement.
It would be better (cache efficient) if we would load one SoA directly before its first usage and unload after the last. However, then this needs to be done by the traversal either through some counter or smart reasoning about the traversal order.